background_remover/requirements.txt
rembg>=2.0.62
Pillow>=10.0.1
click>=8.1.7
tqdm>=4.66.1
onnxruntime>=1.15.0
